rake cron change

+++ lib/tasks/cron.rake
@@ -1,8 +1,10 @@
 task :cron => :environment do
   Event.all.each do |event|
     if event.max_timestamp < Time.now.to_i
-      event.images.destroy_all
-      event.fill_images
+      if event.images.count < 3
+        event.images.destroy_all
+        event.fill_images
+      end
     end
   end
 end
